Style: PR 피드백 수정
- 함수 표현식으로 변경
- Footer 컴포넌트 index.tsx에 구현
